+++ AsBuiltReport.Microsoft.Azure/AsBuiltReport.Microsoft.Azure.psm1
@@ -1,6 +1,6 @@
 # Get public and private function definition files and dot source them
 $Public = @(Get-ChildItem -Path $PSScriptRoot\Src\Public\*.ps1 -ErrorAction SilentlyContinue)
-$Private = @(Get-ChildItem -Path $PSScriptRoot\Src\Private\*.ps1 -ErrorAction SilentlyContinue)
+$Private = @(Get-ChildItem -Path $PSScriptRoot\Src\Private\*.ps1 -Recurse -ErrorAction SilentlyContinue)
 
 foreach ($Module in @($Public + $Private)) {
     try {